// src/app/module-sygmak/models/situation-decadaire.ts
export class SituationDecadaire {
    ideCuoCod(ideCuoCod: any, decRefYer: any, decCod: any, ideAstSer: any, ideAstNbr: any) {
      throw new Error('Method not implemented.');
    }

    repertoire: string ="";
    regime: string ="";
    liquidation: string ="";
    valeur: string ="";
    litrage15: string ="";
    droitstaxes: string ="";
    enregistrement: string ="";
    dateLiquidation: string ="";
    bureau: string ="";
    dateEnregistrement: string ="";
    gds_dsc: string ="";


}

/*
export class SituationDecadaire {
    // Propriétés pour l'affichage des données décadaires
    repertoire: string = "";
    regime: string = "";
    liquidation: string = "";
    valeur: string = "";
    litrage15: string = "";
    droitstaxes: string = "";
    enregistrement: string = "";
    dateLiquidation: string = "";
    bureau: string = "";
    dateEnregistrement: string = "";
    gds_dsc: string = "";
    
    // Propriétés pour l'impression des quittances
    ideCuoCod: string = ""; // Code du bureau de douane
    decRefYer: number = 0; // Année de référence
    decCod: string = "";  // Code de déclaration
    ideAstSer: string = ""; // Série d'assessment
    ideAstNbr: string = ""; // Numéro d'assessment
} */